refactor: parser.rs + to_string
